feat: add function to update status

+++ web/src/helpers/localStorage.jsx
@@ -4,7 +4,7 @@ export const addToReadingList = (book) => {
   const isBookAlreadyAdded = existingList.some((list) => list.id === book.id);
 
   if (!isBookAlreadyAdded) {
-    const bookToAdd = { ...book, status: 'Not Started' };
+    const bookToAdd = { ...book, status: 'Unread' };
     existingList.push(bookToAdd);
 
     localStorage.setItem('@readingList', JSON.stringify(existingList));
@@ -21,6 +21,22 @@ export const removeFromReadingList = (bookId) => {
 
   if (bookIndex !== -1) {
     existingList.splice(bookIndex, 1);
+
+    localStorage.setItem('@readingList', JSON.stringify(existingList));
+
+    return true;
+  } else {
+    return false;
+  }
+};
+
+export const updateBookStatusById = (bookId, newStatus) => {
+  const existingList = JSON.parse(localStorage.getItem('@readingList')) || [];
+
+  const bookIndex = existingList.findIndex((book) => book.id === bookId);
+
+  if (bookIndex !== -1 && existingList[bookIndex].status !== newStatus) {
+    existingList[bookIndex].status = newStatus;
 
     localStorage.setItem('@readingList', JSON.stringify(existingList));
